ai_studio/requirements.txt
google-genai
pdfplumber
python-pptx
gTTS
fastapi
jinja2
faster-whisper
torch
mutagen
